Add macOS patch, fix static analysis
Improvements on PR #731
